multi_agent/requirements-train.txt
# Minimal training requirements (install PyTorch via conda for GPU; see environment.cluster.yml)

ray[rllib]>=2.7,<3.0
gymnasium==0.28.1
pettingzoo==1.23.1
supersuit==3.8.0
stable-baselines3==2.0.0
airsim==1.8.1

numpy==1.25.0
pandas==2.0.3
scipy==1.11.1
matplotlib==3.7.1
opencv-contrib-python==4.8.0.74
pillow==10.0.0
pyyaml==6.0
cloudpickle==2.2.1
psutil==5.9.5
rich==13.4.2
tqdm==4.65.0
shimmy==0.2.1
tensorboard==2.12.3

# AirSim RPC dependencies
msgpack==1.0.7
msgpack-rpc-python==0.4.1
networkx==3.1

# NOTE: Do NOT install torch/torchvision/torchaudio via pip if you need GPU.
#       Prefer conda: `conda install -c pytorch -c nvidia pytorch torchvision torchaudio pytorch-cuda=11.7`.
